New example for subgroup analyses based on the total hip replacement example from the 'Decision Modeling for Health Economic Evaluation' textbook. In process, allow option for one-time 'starting' state values

+++ inst/include/hesim/statmods/obs_index.h
@@ -271,7 +271,7 @@ public:
   /** 
    * Get the patient weight given the current indices
    */    
-  int get_patient_wt(){
+  double get_patient_wt(){
     return patient_wt_vec_[index_];
   }
